requirements.txt
panda-gym>=3.0.0
pybullet>=3.2.0
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
Pillow>=8.0.0

